csv_loader/requirements.txt
google-cloud-bigquery==2.2.0
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
